Rename `collectors/helpers` to `collectors/util`
Change made in order to be consistent with the naming from `tests/`
where directories named `helpers` cannot be used as they will be
ignored by `ava`.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

From https://github.com/avajs/ava/tree/c01ac05ef610c87f5c264b1a3eb80f89b4fc0832#cli:

 " Directories named `fixtures`, `helpers` and `node_modules` are
   always ignored. So are files starting with _ which allows you to
   place helpers in the same directory as your test files. ""
