+++ src/lib/collectors/util/normalize-headers.ts
