+++ src/lib/collectors/util/requester.ts
@@ -14,7 +14,7 @@ import * as iconv from 'iconv-lite';
 
 import { debug as d } from '../../util/debug';
 import { INetworkData } from '../../interfaces'; //eslint-disable-line
-import { RedirectManager } from '../helpers/redirects';
+import { RedirectManager } from './redirects';
 
 const debug = d(__filename);
